lang_example/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
langchain>=0.1.0
sqlalchemy>=2.0.0
pymysql>=1.0.0
redis>=4.0.0
hydra-core>=1.3.0
pydantic>=2.0.0
passlib>=1.7.4
bcrypt>=4.0.0
python-jose>=3.3.0